refactor(bin-designer): address svgParser split review comments
Three review-driven fixes — two from Greptile, one from Copilot:

- Greptile P2: circular type import. svgTransform was importing ViewBox
  from svgParser while svgParser imports Matrix from svgTransform.
  Move ViewBox to ./types (alongside ParsedCutoutSpec, SvgImportError),
  re-export from svgParser for source-tree compatibility, and update
  svgConvertShapes + svgConvertPath to import directly from ./types.
  Cycle eliminated.

- Greptile P2: mutable IDENTITY constant. Wrap with Object.freeze() so
  consumers can't accidentally mutate the shared identity matrix
  reference. Cast back to Matrix since freeze widens the result.

- Copilot P2 (real bug): numAttr was using `parseFloat(...) || fallback`,
  which treats an explicit `0` as falsy and substitutes the fallback.
  An SVG with `<rect x="0">` would have its position silently moved.
  Switched to `Number.isFinite()` check so 0 is preserved and only
  missing/invalid attributes fall back.
